feat: 예약이 존재하는 테마를 삭제할 때 예외 발생

+++ src/main/java/nextstep/service/ThemeService.java
@@ -1,8 +1,10 @@
 package nextstep.service;
 
+import nextstep.dao.ReservationDAO;
 import nextstep.dao.ThemeDAO;
 import nextstep.domain.Theme;
 import nextstep.exceptions.DataNotExistException;
+import nextstep.exceptions.JobNotAllowedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,9 +12,11 @@ import java.util.List;
 @Service
 public class ThemeService {
     private final ThemeDAO themeDAO;
+    private final ReservationDAO reservationDAO;
 
-    public ThemeService(ThemeDAO themeJdbcTemplateDAO) {
+    public ThemeService(ThemeDAO themeJdbcTemplateDAO, ReservationDAO reservationJdbcTemplateDAO) {
         this.themeDAO = themeJdbcTemplateDAO;
+        this.reservationDAO = reservationJdbcTemplateDAO;
     }
 
     public Long saveTheme(Theme theme) {
@@ -24,6 +28,9 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservationDAO.existsByThemeId(id)) {
+            throw new JobNotAllowedException("이 테마를 참조하는 예약이 존재하여 테마를 삭제할 수 없습니다.");
+        }
         if (themeDAO.deleteById(id) == 0) {
             throw new DataNotExistException("테마가 존재하지 않습니다.");
         }
